+++ test/libsolidity/syntaxTests/denominations/szabo_finney_identifiers.sol
@@ -1,0 +1,6 @@
+contract C {
+  function f(uint finney) public pure returns (uint szabo) {
+    // These used to be denominations.
+    szabo = finney;
+  }
+}