+++ test/libsolidity/syntaxTests/parsing/literal_constants_with_ether_subdenominations.sol
@@ -2,15 +2,11 @@ contract C {
     function f() public
     {
          a = 1 wei;
-         b = 2 szabo;
-         c = 3 finney;
-         d = 4 ether;
-         e = 5 gwei;
+         b = 2 ether;
+         c = 3 gwei;
     }
     uint256 a;
     uint256 b;
     uint256 c;
-    uint256 d;
-    uint256 e;
 }
 // ----